requirements.txt
streamlit
tensorflow-cpu
opencv-python-headless
numpy<2.0
Pillow
pandas